+++ classes/alias.js
@@ -5,13 +5,14 @@ const abslog = require('abslog');
 const { join } = require('path');
 const { schemas, validators } = require('@eik/common');
 const { request } = require('../utils/http');
+const { typeSlug } = require('../utils');
 
 module.exports = class Alias {
     constructor({ logger, server, token, type, name, version, alias } = {}) {
         this.log = abslog(logger);
         this.server = server;
         this.token = token;
-        this.type = type;
+        this.type = typeSlug(type);
         this.name = name;
         this.alias = alias;
         this.version = version;
@@ -29,25 +30,26 @@ module.exports = class Alias {
             org: '',
             integrity: '',
         };
-        
+
         this.log.debug('Validating command input');
         schemas.assert.server(this.server);
         schemas.assert.name(this.name);
         schemas.assert.version(this.version);
         validators.type(this.type);
         validators.alias(this.alias);
-        assert(this.token && typeof this.token === 'string', `Parameter "token" is not valid`);
+        assert(
+            this.token && typeof this.token === 'string',
+            `Parameter "token" is not valid`,
+        );
 
-        this.log.debug(`Requesting creation of ${this.type} alias "v${this.alias}" for ${this.name} v${this.version}`);
+        this.log.debug(
+            `Requesting creation of ${this.type} alias "v${this.alias}" for ${this.name} v${this.version}`,
+        );
         try {
             const { message } = await request({
                 host: this.server,
                 method: 'PUT',
-                pathname: join(
-                    this.type,
-                    this.name,
-                    `v${this.alias}`,
-                ),
+                pathname: join(this.type, this.name, `v${this.alias}`),
                 data: { version: this.version },
                 token: this.token,
             });
@@ -63,17 +65,15 @@ module.exports = class Alias {
             let status = err.statusCode;
 
             if (status === 409) {
-                this.log.debug('Alias already exists on server, performing update');
+                this.log.debug(
+                    'Alias already exists on server, performing update',
+                );
 
                 try {
                     const { message: msg } = await request({
                         host: this.server,
                         method: 'POST',
-                        pathname: join(
-                            this.type,
-                            this.name,
-                            `v${this.alias}`,
-                        ),
+                        pathname: join(this.type, this.name, `v${this.alias}`),
                         data: { version: this.version },
                         token: this.token,
                     });
@@ -111,7 +111,9 @@ module.exports = class Alias {
                         'Client attempted to send an unsupported file format to server',
                     );
                 case 502:
-                    throw new Error('Server was unable to write file to storage');
+                    throw new Error(
+                        'Server was unable to write file to storage',
+                    );
                 default:
                     throw new Error('Server failure');
             }