+++ commands/publish.js
@@ -8,19 +8,20 @@ const {
     helpers: { configStore },
 } = require('@eik/common');
 const PublishPackage = require('../classes/publish/package/index');
-const { logger, getDefaults, getCWD } = require('../utils');
+const {
+    logger,
+    getDefaults,
+    getCWD,
+    typeSlug,
+    typeTitle,
+} = require('../utils');
 const { Artifact } = require('../formatters');
 
-exports.command = 'package';
+exports.command = 'publish';
 
-exports.aliases = ['pkg', 'publish'];
+exports.aliases = ['pkg', 'package', 'pub'];
 
-exports.describe = `Publish an app package to an Eik server by a given semver level.
-    Bundles and publishes JavaScript and CSS at given local paths creating a new version based off the previous version and the given semver level.
-    Specifying semver level is optional and defaults to "patch", specifying "major" locks version to the given semver major.
-    If a package has never previously been published, the first version generated will be equal to specified major version or 1.0.0 if no major is specified.
-    Local paths to asset files can be defined in an "eik.json" file using "js.input" and "css.input" fields or can be provided directly to the CLI command using the flags --js and --css.
-    URLs to import maps can be provided to map "bare" imports found in asset files, to do so either use the field "map" in "eik.json" or the --map CLI flag.`;
+exports.describe = `Publish an app package to an Eik server. Reads configuration from eik.json or package.json files. See https://eik.dev for more details.`;
 
 exports.builder = (yargs) => {
     const cwd = getCWD();
@@ -45,14 +46,8 @@ exports.builder = (yargs) => {
             default: false,
             type: 'boolean',
         },
-        npm: {
-            describe: 'Assigns package to the NPM namespace',
-            default: false,
-            type: 'boolean',
-        },
         token: {
-            describe:
-                `Provide a jwt token to be used to authenticate with the Eik server.
+            describe: `Provide a jwt token to be used to authenticate with the Eik server.
                 Automatically determined if authenticated (via eik login)`,
             type: 'string',
             alias: 't',
@@ -61,41 +56,49 @@ exports.builder = (yargs) => {
 
     yargs.default('token', defaults.token, defaults.token ? '######' : '');
 
-    yargs.example(`eik package`);
     yargs.example(`eik publish`);
-    yargs.example(`eik publish --dry-run`);
+    yargs.example(`eik package`);
+    yargs.example(`eik pub --dry-run`);
     yargs.example(`eik pkg --token ######`);
     yargs.example(`eik pkg --debug`);
 };
 
 exports.handler = async (argv) => {
     const spinner = ora({ stream: process.stdout }).start('working...');
-    const { debug, dryRun, cwd, token, npm } = argv;
+    const { debug, dryRun, cwd, token } = argv;
     const config = configStore.findInDirectory(cwd);
-    const {name, server, map, version, out} = config;
+    const { name, server, version, type, map, out, files } = config;
+
+    if (type === 'map') {
+        spinner.warn(
+            '"type" is set to "map", which is not supported by the publish command. Please use the "eik map" command instead',
+        );
+        process.stdout.write('\n');
+        process.exit(0);
+    }
 
     try {
-        const options = { 
+        const options = {
             logger: logger(spinner, debug),
-            name,
-            server,
-            map: Array.isArray(map) ? map : [map],
-            config,
-            version,
             cwd,
             token,
             dryRun,
             debug,
+            name,
+            server,
+            version,
+            type,
+            map,
             out,
-            npm,
+            files,
         };
-
-        const type = npm ? 'npm' : 'pkg';
 
         const publish = await new PublishPackage(options).run();
 
         if (!publish) {
-            spinner.warn('Version in eik.json has not changed since last publish, publishing is not necessary');
+            spinner.warn(
+                'Version in eik.json has not changed since last publish, publishing is not necessary',
+            );
             process.stdout.write('\n');
             process.exit(0);
         }
@@ -103,16 +106,16 @@ exports.handler = async (argv) => {
         const { files: fls } = publish;
 
         if (!dryRun) {
-            let url = new URL(join(type, name), server);
+            let url = new URL(join(typeSlug(type), name), server);
             let res = await fetch(url);
             const pkgMeta = await res.json();
 
-            url = new URL(join(type, name, version), server);
+            url = new URL(join(typeSlug(type), name, version), server);
             res = await fetch(url);
             const pkgVersionMeta = await res.json();
 
             const artifact = new Artifact(pkgMeta);
-            artifact.versions = [ pkgVersionMeta ];
+            artifact.versions = [pkgVersionMeta];
 
             spinner.text = '';
             spinner.stopAndPersist();
@@ -123,14 +126,26 @@ exports.handler = async (argv) => {
             spinner.text = '';
             spinner.stopAndPersist();
 
-            process.stdout.write(`:: ${chalk.bgYellow.white.bold(npm ? ' NPM ' : ' PACKAGE ')} > ${chalk.green(name)} | ${chalk.bold('dry run')}`);
+            process.stdout.write(
+                `:: ${chalk.bgYellow.white.bold(
+                    typeTitle(type),
+                )} > ${chalk.green(name)} | ${chalk.bold('dry run')}`,
+            );
             process.stdout.write('\n\n');
             process.stdout.write('   files (local temporary):\n');
             for (const file of fls) {
-                process.stdout.write(`   - ${chalk.bold('type')}: ${file.type}\n`);
-                process.stdout.write(`     ${chalk.bold('path')}: ${file.pathname}\n\n`);
+                process.stdout.write(
+                    `   - ${chalk.bold('type')}: ${file.type}\n`,
+                );
+                process.stdout.write(
+                    `     ${chalk.bold('path')}: ${file.pathname}\n\n`,
+                );
             }
-            process.stdout.write(`   ${chalk.bold('No files were published to remote server')}\n\n`);
+            process.stdout.write(
+                `   ${chalk.bold(
+                    'No files were published to remote server',
+                )}\n\n`,
+            );
         }
     } catch (err) {
         spinner.warn(err.message);