+++ formatters/file.js
@@ -5,12 +5,12 @@
 
 'use strict';
 
-const {join} = require('path');
+const { join } = require('path');
 const chalk = require('chalk');
 
 function readableBytes(bytes) {
     const i = Math.floor(Math.log(bytes) / Math.log(1024)),
-    sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
+        sizes = ['B', 'KB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
     return (bytes / Math.pow(1024, i)).toFixed(2) * 1 + ' ' + sizes[i];
 }