+++ test/integration/meta.test.js
@@ -27,28 +27,26 @@ beforeEach(async (done, t) => {
     const address = await server.listen();
     const folder = await fs.mkdtemp(join(os.tmpdir(), basename(__filename)));
     const eik = join(__dirname, '../../index.js');
-    
-    const token = await new cli.Login({
+
+    const token = await cli.login({
         server: address,
         key: 'change_me',
-    }).run();
+    });
 
     const assets = {
         name: 'scroll-into-view-if-needed',
         version: '2.2.24',
+        type: 'npm',
         server: address,
         files: {
-            './index.js': join(__dirname, './../fixtures/client.js'),
-            './index.css': join(__dirname, './../fixtures/styles.css'),
+            'index.js': join(__dirname, './../fixtures/client.js'),
+            'index.css': join(__dirname, './../fixtures/styles.css'),
         },
     };
 
-    await fs.writeFile(
-        join(folder, 'eik.json'),
-        JSON.stringify(assets),
-    );
+    await fs.writeFile(join(folder, 'eik.json'), JSON.stringify(assets));
 
-    const cmd = `${eik} package --token ${token} --cwd ${folder} --npm`;
+    const cmd = `${eik} package --token ${token} --cwd ${folder}`;
     await exec(cmd);
 
     t.context.server = server;
@@ -63,34 +61,12 @@ afterEach(async (done, t) => {
     done();
 });
 
-test('eik meta --server : no eik.json', async (t) => {
-    const eik = join(__dirname, '../../index.js');
-    const cmd = `${eik} meta scroll-into-view-if-needed --server ${t.context.address}`;
-
-    const { error, stdout } = await exec(cmd);
-
-    t.notOk(error);
-    t.match(stdout, '::');
-    t.match(stdout, 'NPM');
-    t.match(stdout, 'scroll-into-view-if-needed');
-    t.end();
-});
-
-test('eik meta : details provided by eik.json', async (t) => {
-    const assets = {
-        name: 'test-app',
-        server: t.context.address,
-    };
-    await fs.writeFile(
-        join(t.context.folder, 'eik.json'),
-        JSON.stringify(assets),
-    );
-
+test('eik meta', async (t) => {
     const eik = join(__dirname, '../../index.js');
     const cmd = `${eik} meta scroll-into-view-if-needed --cwd ${t.context.folder}`;
 
     const { error, stdout } = await exec(cmd);
-    
+
     t.notOk(error);
     t.match(stdout, '::');
     t.match(stdout, 'NPM');