+++ test/publish.map.test.js
@@ -18,16 +18,15 @@ beforeEach(async (done, t) => {
     const service = new EikService({ customSink: memSink });
     server.register(service.api());
     const address = await server.listen();
-    
-    const login = new cli.Login({
+
+    const token = await cli.login({
         server: address,
         key: 'change_me',
     });
-    const token = await login.run();
-    
+
     const cwd = await fs.mkdtemp(join(os.tmpdir(), basename(__filename)));
 
-    t.context.server = server
+    t.context.server = server;
     t.context.address = address;
     t.context.token = token;
     t.context.cwd = cwd;
@@ -39,11 +38,11 @@ afterEach(async (done, t) => {
     done();
 });
 
-test('Uploading import map to an asset server', async t => {
+test('Uploading import map to an asset server', async (t) => {
     const { address, token, cwd } = t.context;
     const l = mockLogger();
 
-    const publishMap = new cli.publish.Map({
+    const result = await cli.map({
         logger: l.logger,
         cwd,
         server: address,
@@ -54,13 +53,16 @@ test('Uploading import map to an asset server', async t => {
         token,
     });
 
-    const result = await publishMap.run();
-    t.same(result, {
-        name: 'my-map',
-        version: '1.0.0',
-        server: address,
-        type: 'map',
-    }, 'Command should return an object');
+    t.same(
+        result,
+        {
+            name: 'my-map',
+            version: '1.0.0',
+            server: address,
+            type: 'map',
+        },
+        'Command should return an object',
+    );
     t.match(
         l.logs.debug,
         'Uploading import map "my-map" version "1.0.0" to asset server',