+++ utils/hash/compare.js
@@ -7,7 +7,7 @@
  * @param {string} hash2 - second hash string to compare
  *
  * @returns {boolean}
- * 
+ *
  * @example hash.compare('a1b22c23d24e25f4g33a123b23c34', 'a1b22c23d24e25f4g33a123b23c34');
  */
 module.exports = (hash1, hash2) => hash1 === hash2;