+++ utils/index.js
@@ -5,5 +5,7 @@ const {
 } = require('@eik/common');
 const logger = require('./logger');
 const getCWD = require('./get-cwd');
+const typeSlug = require('./type-slug');
+const typeTitle = require('./type-title');
 
-module.exports = { logger, getDefaults, getCWD, files };
+module.exports = { logger, getDefaults, getCWD, files, typeSlug, typeTitle };